Test tarballs that are similar to release tarballs
Call `make dist` outside of the test VMs instead of calling autogen.sh
inside. Adjust the debian and rpm packaging to reflect that.

Tests are now running on tarballs that are more similar to release
tarballs and test packages are more similar to packages that
distributions ship.

The system running the tests must now also have build dependencies
installed.

make-source, the script that prepares the tarball, tries not to
interfere too much with existing builds. It configures the package in
tmp-dist/ without setting up the redirect Makefile. However, due to
limitations in autotools, it must use an existing in-tree configuration.

Closes #4178

+++ autogen.sh
@@ -36,7 +36,6 @@ olddir=$(pwd)
 cd $srcdir
 
 npm install # see package.json
-cp package.json node_modules # for tools/make-source
 find node_modules -name test | xargs rm -rf
 
 rm -rf autom4te.cache
@@ -68,7 +67,7 @@ rm -f $srcdir/Makefile
 $srcdir/configure --enable-maintainer-mode ${AUTOGEN_CONFIGURE_ARGS:-} "$@" || exit $?
 
 # Put a redirect makefile here
-if [ ! -f $srcdir/Makefile ]; then
+if [ -z "${NOREDIRECTMAKEFILE:-}" -a ! -f $srcdir/Makefile ]; then
     cat $srcdir/tools/Makefile.redirect > $srcdir/Makefile
     printf "\nREDIRECT = %s\n" "$(realpath $olddir)" >> $srcdir/Makefile
 fi
